pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nautobot-ipsec-builder"
version = "0.1.0"
description = "Nautobot app to build IPsec tunnels on Cisco IOS-XE devices"
authors = [{ name = "NRTC Ops" }]
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "nautobot>=3.0.0,<4.0.0",
    "netmiko>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "black",
    "flake8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["nautobot_ipsec_builder*"]

[tool.setuptools.package-data]
"nautobot_ipsec_builder" = [
    "templates/**/*.html",
    "static/**/*",
]
